requirements.txt
email-validator>=2.2.0
flask>=3.1.1
flask-sqlalchemy>=3.1.1
google-cloud-texttospeech>=2.27.0
google-genai>=1.18.0
google-generativeai>=0.8.5
gunicorn>=23.0.0
psycopg2-binary>=2.9.10
python-dotenv>=1.1.0
sift-stack-py>=0.6.1
